feat(substructures): #CO-1796 add possibility to view substructures of a master structure

+++ src/main/java/net/atos/entng/support/services/impl/TicketServiceSqlImpl.java
@@ -195,7 +195,7 @@ public class TicketServiceSqlImpl extends SqlCrudService implements TicketServic
 
 	@Override
 	public Future<JsonArray> listTickets(UserInfos user, Integer page, List<String> statuses, List<String> applicants, String school_id,
-							String sortBy, String order, Integer nbTicketsPerPage, JsonArray orderedStructures) {
+							String sortBy, String order, Integer nbTicketsPerPage, JsonArray orderedStructures, JsonObject structureChildrens) {
 		Promise<JsonArray> promise = Promise.promise();
 		StringBuilder query = new StringBuilder();
 		query.append("SELECT t.*, u.username AS owner_name, ")
@@ -220,13 +220,23 @@ public class TicketServiceSqlImpl extends SqlCrudService implements TicketServic
 		if (adminLocal != null) {
 			List<String> scopesList = adminLocal.getScope();
 			if(scopesList != null && !scopesList.isEmpty()) {
-				query.append(" WHERE ((t.school_id IN (");
-				for (String scope : scopesList) {
-					query.append("?,");
-					values.add(scope);
+				if (school_id.equals("*")){
+					query.append(" AND ((t.school_id IN (");
+					for (String scope : scopesList) {
+						query.append("?,");
+						values.add(scope);
+					}
+					query.deleteCharAt(query.length() - 1);
+					query.append(")");
+				}else if(scopesList.contains(school_id)){
+					query.append(" AND t.school_id IN ");
+					List<String> listIdStructure = structureChildrens.getJsonArray(Ticket.STRUCTUREIDS).stream()
+							.filter(String.class::isInstance)
+							.map(Object::toString)
+							.collect(Collectors.toList());
+					query.append(Sql.listPrepared(listIdStructure));
+					values.addAll(new JsonArray(listIdStructure));
 				}
-				query.deleteCharAt(query.length() - 1);
-				query.append(")");
 
 				if (oneApplicant) {
 					query.append(" AND t.owner").append(applicantIsMe?"=":"!=").append("?");
@@ -242,8 +252,8 @@ public class TicketServiceSqlImpl extends SqlCrudService implements TicketServic
 				query.append(")");
 			}
 		}
-		else {
-            query.append(" WHERE t.school_id IN ").append(Sql.listPrepared(user.getStructures())).append(" ");
+		else if (school_id.equals("*")){
+            query.append(" AND t.school_id IN ").append(Sql.listPrepared(user.getStructures()));
             values.addAll(new JsonArray(user.getStructures()));
 			if (oneApplicant) {
 				query.append(" AND t.owner").append(applicantIsMe?"=":"!=").append("?");
@@ -251,7 +261,7 @@ public class TicketServiceSqlImpl extends SqlCrudService implements TicketServic
 			}
         }
 
-		if (statuses.size() > 0 && statuses.size() < 4) {
+		if (statuses.size() > 0 && statuses.size() < 5) {
 			query.append(" AND (t.status IN (");
 			for (String status : statuses) {
 				query.append("?,");
@@ -262,8 +272,13 @@ public class TicketServiceSqlImpl extends SqlCrudService implements TicketServic
 		}
 
 		if (!school_id.equals("*")) {
-			query.append(" AND t.school_id = ?");
-			values.add(school_id);
+			query.append(" AND t.school_id IN ");
+			List<String> listIdStructure = structureChildrens.getJsonArray(Ticket.STRUCTUREIDS).stream()
+					.filter(String.class::isInstance)
+					.map(Object::toString)
+					.collect(Collectors.toList());
+			query.append(Sql.listPrepared(listIdStructure));
+			values.addAll(new JsonArray(listIdStructure));
 		}
 
 
@@ -291,7 +306,7 @@ public class TicketServiceSqlImpl extends SqlCrudService implements TicketServic
 			values.add(nbTicketsPerPage);
 		}
 
-		sql.prepared(query.toString(), values, validResultHandler(PromiseHelper.handler(promise)));
+		sql.prepared(query.toString().replaceFirst("AND", "WHERE"), values, validResultHandler(PromiseHelper.handler(promise)));
 		return promise.future();
 	}
